Whole bunch of stuff, basically stable
I was going to rewrite the last commit into this
one, but I wasn't expecting to change as much
as I did in this one, so here we are.

New thresholds to account for questionable calibration
on the board.

Set the solar voltage threshold based on crystal's
recommendation, to be tested for real-life issues

Make things work as expected

Add status LED for temperature indication, for now